requirements.txt
requests==2.32.5
beautifulsoup4==4.13.5
pandas==2.3.2
tqdm==4.67.1
python-dotenv==1.0.0
